% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map the data for a given state for a given year}
\source{
The data used is from the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System (FARS), which is a nationwide census providing the
American public yearly data regarding fatal injuries suffered in motor vehicle
traffic crashes.
}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric value from 1 to 51 that represents each of the states and DC}

\item{year}{A specific year in YYYY format}
}
\value{
A geographic map of the state with points representing fatalities during that year
}
\description{
This function takes a state number and a given year in YYYY format
and displays a geogrphical map of fatalities accross the state
for that year
}
\examples{
\dontrun{
fars_map_state(24, 2014)
fars_map_state(51, 2015)
}

}

