% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read data for a list of years}
\source{
The data used is from the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System (FARS), which is a nationwide census providing the
American public yearly data regarding fatal injuries suffered in motor vehicle
traffic crashes.
}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years in YYYY format}
}
\value{
A list of data tables for each year.
     Returns an error message "invalid year YYYY" when an error is encountered.
}
\description{
This function takes a list of years in YYYY format and reads the FARS data
for each year
}
\examples{
fars_read_years(2013)
fars_read_years(c(2014, 2015))
fars_read_years(2013:2015)

}

