% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the data for each year}
\source{
The data used is from the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System (FARS), which is a nationwide census providing the
American public yearly data regarding fatal injuries suffered in motor vehicle
traffic crashes.
}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of years in YYYY format}
}
\value{
A table of the total fatalities per month in each year
}
\description{
This function takes a list of years and returns a table of the total number
of fatalities per month for each year
}
\examples{
fars_summarize_years(2013)
fars_sumarize_years(c(2014, 2015))
fars_summarize_years(2013:2015)

}

